Fix SwiftLint violation for "file_length" rule by moving functions for adjusting scroll view when keyboard appears to its own file.
