+++ .swiftlint.yml
@@ -12,7 +12,6 @@ disabled_rules:
   - unused_closure_parameter
   - todo
   - cyclomatic_complexity
-  - file_length
   - function_parameter_count
   - function_body_length
