+++ FourSix Coffee Timer/Controllers/Notes/NoteDetails/RegisterKeyboardNotifications+NoteDetailsVC.swift
@@ -1,0 +1,38 @@
+//
+//  RegisterKeyboardNotifications+NoteDetailsVC.swift
+//  FourSix Coffee Timer
+//
+//  Created by Jon Duenas on 5/17/21.
+//  Copyright © 2021 Jon Duenas. All rights reserved.
+//
+
+import UIKit
+
+extension NoteDetailsVC {
+    func registerKeyboardNotifications() {
+        NotificationCenter.default.addObserver(self, selector: #selector(adjustForKeyboard), name: UIResponder.keyboardWillHideNotification, object: nil)
+        NotificationCenter.default.addObserver(self, selector: #selector(adjustForKeyboard), name: UIResponder.keyboardWillChangeFrameNotification, object: nil)
+    }
+    
+    @objc func adjustForKeyboard(notification: Notification) {
+        guard let keyboardValue = notification.userInfo?[UIResponder.keyboardFrameEndUserInfoKey] as? NSValue else { return }
+        
+        let keyboardScreenEndFrame = keyboardValue.cgRectValue
+        let keyboardViewEndFrame = view.convert(keyboardScreenEndFrame, from: view.window)
+        
+        if notification.name == UIResponder.keyboardWillHideNotification {
+            scrollView.contentInset = .zero
+        } else {
+            scrollView.contentInset = UIEdgeInsets(top: 0,
+                                                   left: 0,
+                                                   bottom: keyboardViewEndFrame.height - view.safeAreaInsets.bottom - (tabBarController?.tabBar.frame.height ?? 0) + 10,
+                                                   right: 0)
+        }
+        
+        scrollView.scrollIndicatorInsets = scrollView.contentInset
+        
+        if notesTextView.isFirstResponder {
+            scrollView.scrollRectToVisible(notesTextView.frame, animated: true)
+        }
+    }
+}
